+++ src/@next/components/organisms/PaymentGatewaysList/PaymentGatewaysList.tsx
@@ -7,6 +7,7 @@ import {
   BraintreePaymentGateway,
   DummyPaymentGateway,
   StripePaymentGateway,
+  AdyenPaymentGateway,
 } from "..";
 import * as S from "./styles";
 import { IProps } from "./types";
@@ -22,6 +23,8 @@ const PaymentGatewaysList: React.FC<IProps> = ({
   formRef,
   formId,
   processPayment,
+  submitPayment,
+  submitPaymentSuccess,
   errors,
   onError,
 }: IProps) => {
@@ -122,6 +125,41 @@ const PaymentGatewaysList: React.FC<IProps> = ({
                     processPayment={(token, cardData) =>
                       processPayment(id, token, cardData)
                     }
+                    errors={errors}
+                    onError={onError}
+                  />
+                )}
+              </div>
+            );
+
+          case PROVIDERS.ADYEN.label:
+            return (
+              <div key={index}>
+                <S.Tile checked={checked}>
+                  <Radio
+                    data-test="checkoutPaymentGatewayAdyenInput"
+                    name="payment-method"
+                    value="adyen"
+                    checked={checked}
+                    onChange={() =>
+                      selectPaymentGateway && selectPaymentGateway(id)
+                    }
+                    customLabel
+                  >
+                    <span data-test="checkoutPaymentGatewayAdyenName">
+                      {name}
+                    </span>
+                  </Radio>
+                </S.Tile>
+                {checked && (
+                  <AdyenPaymentGateway
+                    config={config}
+                    formRef={formRef}
+                    scriptConfig={PROVIDERS.ADYEN.script}
+                    styleConfig={PROVIDERS.ADYEN.style}
+                    processPayment={() => processPayment(id)}
+                    submitPayment={submitPayment}
+                    submitPaymentSuccess={submitPaymentSuccess}
                     errors={errors}
                     onError={onError}
                   />