+++ src/@next/components/organisms/PaymentGatewaysList/types.ts
@@ -1,4 +1,5 @@
 import { ICardData, IFormError, IPaymentGateway } from "@types";
+import { CompleteCheckout_checkoutComplete_order } from "@saleor/sdk/lib/mutations/gqlTypes/CompleteCheckout";
 
 export interface IProps {
   /**
@@ -34,8 +35,15 @@ export interface IProps {
    */
   processPayment: (
     gateway: string,
-    token: string,
+    token?: string,
     cardData?: ICardData
+  ) => void;
+  submitPayment: (data: {
+    confirmationData: any;
+    confirmationNeeded: boolean;
+  }) => Promise<any>;
+  submitPaymentSuccess: (
+    order?: CompleteCheckout_checkoutComplete_order
   ) => void;
   /**
    * Method called when gateway error occured.